new stuff, settings base added, scoreboard added, eco system addon

+++ src/main/java/me/miskynet/customGamemode/custom/item/shop/ShopItem.java
@@ -17,13 +17,13 @@ import java.util.List;
 
 public class ShopItem extends Item {
 
-    int buyPrice;
-    int sellPrice;
+    Double buyPrice;
+    Double sellPrice;
 
     private static final NamespacedKey buyKey = new NamespacedKey(Main.getInstance(), "buy_price");
     private static final NamespacedKey sellKey = new NamespacedKey(Main.getInstance(), "sell_price");
 
-    public ShopItem(Material material, Integer buyPrice, Integer sellPrice) {
+    public ShopItem(Material material, Double buyPrice, Double sellPrice) {
         super(material, null, new ArrayList<>());
 
         List<Component> lore = new ArrayList<>();
@@ -42,10 +42,10 @@ public class ShopItem extends Item {
         ItemMeta meta = itemStack.getItemMeta();
         var container = meta.getPersistentDataContainer();
 
-        if (container.has(buyKey, PersistentDataType.INTEGER) && container.has(sellKey, PersistentDataType.INTEGER)) {
+        if (container.has(buyKey, PersistentDataType.DOUBLE) && container.has(sellKey, PersistentDataType.DOUBLE)) {
 
-            int buy = container.get(buyKey, PersistentDataType.INTEGER);
-            int sell = container.get(sellKey, PersistentDataType.INTEGER);
+            Double buy = container.get(buyKey, PersistentDataType.DOUBLE);
+            Double sell = container.get(sellKey, PersistentDataType.DOUBLE);
 
             return new ShopItem(itemStack.getType(), buy, sell);
         }
@@ -53,8 +53,8 @@ public class ShopItem extends Item {
         return null;
     }
 
-    public int getBuyPrice() { return buyPrice; }
-    public int getSellPrice() { return sellPrice; }
+    public Double getBuyPrice() { return buyPrice; }
+    public Double getSellPrice() { return sellPrice; }
 
     @Override
     public ItemStack toItemStack() {
@@ -62,8 +62,8 @@ public class ShopItem extends Item {
         ItemMeta meta = itemStack.getItemMeta();
 
         if (meta != null) {
-            meta.getPersistentDataContainer().set(buyKey, PersistentDataType.INTEGER, this.buyPrice);
-            meta.getPersistentDataContainer().set(sellKey, PersistentDataType.INTEGER, this.sellPrice);
+            meta.getPersistentDataContainer().set(buyKey, PersistentDataType.DOUBLE, this.buyPrice);
+            meta.getPersistentDataContainer().set(sellKey, PersistentDataType.DOUBLE, this.sellPrice);
 
             itemStack.setItemMeta(meta);
         }
